moved test to SolrUpdateIntegrationTest
